requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
sqlalchemy>=2.0.35
python-multipart>=0.0.20
python-docx>=1.1.2
pdfplumber>=0.11.4
jinja2>=3.1.4
pydantic>=2.11.0
httpx>=0.27.2
pytest>=8.3.3
pyinstaller==6.18.0
